services/backend/requirements.txt
alembic==0.9.8
aniso8601==3.0.0
bcrypt==3.1.4
cffi==1.11.5
cliapp==1.0.9
click==6.7
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-Redis==0.3.0
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SocketIO==3.3.2
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
pycparser==2.18
PyJWT==1.6.0
PyMySQL==0.9.3
python-dateutil==2.7.0
python-editor==1.0.3
python-engineio==3.5.0
python-socketio==3.1.2
pytz==2018.3
redis==3.2.1
six==1.11.0
SQLAlchemy==1.2.5
Werkzeug==0.14.1
gunicorn
eventlet

opencv-python
Pillow

tensorflow==2.0.0
keras==2.3.1
scikit-learn
imageio
